feat: add csv utils module
